Rework Debian description determination
Now the package description is taken from the first set variable from the
following:

* CPACK_DEBIAN_COMPONENT_<COMP>_DESCRIPTION
* CPACK_COMPONENT_<COMP>_DESCRIPTION
* CPACK_DEBIAN_PACKAGE_DESCRIPTION
* CPACK_PACKAGE_DESCRIPTION_SUMMARY

This is compatible with the upstream logic (which only uses the last
two variables), offers a reasonable default when no Debian-specific
variables are set, and the name of the new variable (the first one)
is consistent with both the second and the third ones.

I had to remove the summary/long description split, because I found
no way to achieve all of the above results while keeping it. It's
easy enough to properly format the description in user code, if need be.